refactor(desktop): split bot-chat settings page into overview and detail components
Refs #1042. BotChatSettingsPage fused overview routing, the detail view,
status fetch/subscribe, and four action lifecycles in one ~960-line
component. Split the two views into sibling components:

- bot-chat-overview.tsx: BotChatOverview (active/available channel rows,
  derived during render from channels + statuses props)
- bot-chat-detail.tsx: BotChatChannelDetail (header/enable switch, runtime
  status + action stack, credential form, WeChat login modals) plus the
  moved BotAllowedUserIdsField / botConnectionLabel / canEnableBotChannel
- bot-chat-shared.tsx: BOT_LABELS, readiness copy, BotBrandLogo,
  botStatusDetail, and the pending-action types shared by all three

The page keeps routing, status fetching, and the action lifecycles and
passes them down as props. Also drops dead code: csvList, the unused
`testing` derivation, and the unused useId/LlmConnection imports.

Contract tests co-migrated to pin the same invariants against the new
structure (identifier renames selected->provider/status in the detail,
props.actionBusy/props.restarting, botDetailMountedRef for the scan-login
continuation, new files added to the settings combined-source helper).
