+++ apps/desktop/src/main/__tests__/renderer-startup-fail-soft-contract.test.ts
@@ -101,8 +101,10 @@ describe('renderer startup fail-soft contract', () => {
       /catch[\s\S]*setStatuses\(null\)/,
       'bot status probe failure must preserve current statuses instead of clearing them',
     );
-    assert.match(botPage, /<Alert variant="error">[\s\S]*<AlertTitle>远程接入状态载入失败<\/AlertTitle>[\s\S]*<AlertDescription>\{statusLoadError\}<\/AlertDescription>/);
-    assert.match(botPage, /<Alert variant="error">[\s\S]*<AlertTitle>运行状态刷新失败<\/AlertTitle>[\s\S]*<AlertDescription>\{statusLoadError\}<\/AlertDescription>/);
+    // #1042: the overview/detail views render the page's statusLoadError
+    // via props after the bot-chat split.
+    assert.match(botPage, /<Alert variant="error">[\s\S]*<AlertTitle>远程接入状态载入失败<\/AlertTitle>[\s\S]*<AlertDescription>\{props\.statusLoadError\}<\/AlertDescription>/);
+    assert.match(botPage, /<Alert variant="error">[\s\S]*<AlertTitle>运行状态刷新失败<\/AlertTitle>[\s\S]*<AlertDescription>\{props\.statusLoadError\}<\/AlertDescription>/);
     assert.match(
       botPage,
       /async function refreshBotStatuses\(\): Promise<boolean> \{[\s\S]*try \{[\s\S]*window\.maka\.settings\.bots\.listStatuses\(\)[\s\S]*setStatuses\(nextStatuses\)[\s\S]*setStatusLoadError\(null\)[\s\S]*return true;[\s\S]*\} catch \(error\) \{[\s\S]*setStatusLoadError\(message\);[\s\S]*toast\.error\('刷新远程接入状态失败', message\)[\s\S]*return false;/,
