+++ apps/desktop/src/main/__tests__/settings-contract-source-helpers.ts
@@ -28,6 +28,9 @@ const sourcePaths = [
   'general-settings-page.tsx',
   'open-gateway-settings-page.tsx',
   'bot-chat-settings-page.tsx',
+  'bot-chat-shared.tsx',
+  'bot-chat-overview.tsx',
+  'bot-chat-detail.tsx',
   'bot-wechat-login.tsx',
   'usage-settings-page.tsx',
   'settings-metric-card.tsx',
